cec: Change CEC_UNREGISTERED_ADDR to 15 and add CEC_INVALID_ADDR
Per "CEC Table 5 Logical Addresses" in the HDMI 1.4 spec, the
unregistered logical address is 15. The spec says this should be used
when a device has no valid physical address (e.g. it's unplugged), or as
a fallback if no type-specific logical addresses are available.

Currently the EC defines CEC_UNREGISTERED_ADDR as 255. This is actually
the invalid logical address, which is used to indicate that no logical
address has been set. It is not defined in the spec, but the kernel
defines it as CEC_LOG_ADDR_INVALID in include/uapi/linux/cec.h and uses
it e.g. when an adapter is unconfigured.

Change CEC_UNREGISTERED_ADDR to 15 and add CEC_INVALID_ADDR = 255.
There are no functional changes.

BUG=b:270507438
TEST=CEC works on dibbi

Change-Id: I9ce8a091e246da356ddea8ac04910c6d46093cc4