ec: Fix logic of asyc flash protect command
The async version of Flash protect command is hung in an EC_RES_ERROR
status in case a previous command returned this status. Correct the
logic flow to avoid sending another crec set flash protect command in
case the status is busy and reset the status otherwise.

BUG=b:116396469
TEST=make buildall -j
TEST= run flashprotect command using the ectool. Force wp state to be
enabled and check ectool command output with enable, disable, and none
configurations.

Change-Id: I2c7b557e6add865d3cd6382927de0a3475c17e73