bazel: Allow separate board name from target name
This adds a new parameter to ec_target, board=..., which when
specified, allows the board name to differ from the target name.

This is useful for targets which exist for both Legacy EC and Zephyr
EC, so we can name the Zephyr target :brya_zephyr, for example.  I
updated BUILD.bazel to show this example.

BUG=b:290816553
TEST=bazel build :brya_zephyr
     bazel build :brya (fails during genvif, as expected ATM)

Change-Id: Id52382e779cd7c7e4c69b8f88646924256eb62ba