bazel: Auto-generate the targets list
Now that we support both Legacy EC and Zephyr EC, the targets list is
becoming difficult to hand-maintain.  Add a script which can
auto-generate the bazel.

In the future, we'll add validation to ensure that all new
targets (and deleted targets) result in the file getting re-generated,
or let the CQ check in that change as well.

BUG=b:290816553
TEST=inspect output
TEST=bazel build :brya_zephyr
TEST=bazel build :arcada_ish

Change-Id: I573b314a4b4c3b97a4429356eb785e5fcadb5b6d