rex: Enable BATTERY_HW_PRESENT_CUSTOM
This enables the battery_hw_present() check which uses a GPIO to
definitively detect the presence of the system battery. This enables
some additional system power-loss avoidance checks where appropriate.

This is particularly relevant because in modern usage
battery_is_present() is more of a logical "battery usable" check rather
than a physical battery presence check.

BUG=b:289092921
TEST='zmake build rex' passes

Change-Id: I1f5889b3012adba157e1da56545d7300c7c1dd02