rex: Enable BATTERY_HW_PRESENT_CUSTOM
This enables the battery_hw_present() check which uses a GPIO to
definitively detect the presence of the system battery. This enables
some additional system power-loss avoidance checks where appropriate.

This is particularly relevant because in modern usage
battery_is_present() is more of a logical "battery usable" check rather
than a physical battery presence check.

BUG=b:289566759

Change-Id: I0cadc6e88a0b4820b321ff95ec0476b02042d43c