charger: Only update battery info when it is present
The charger task will query the battery info if needed. However, if
battery is absent, there is no battery to query its info, so add the
guarding statement via checking is battery is present beforehand.

For smart batteries, querying the info is achieved via i2c transactions,
and if the price of failing the i2c transaction on certain chips are
high (such as it8xxx2), the system performance is greatly impacted, thus
affecting the behavior of the device.

BUG=b:290553696
TEST=(1) Connect the battery and boot, EC is working as before.
     (2) Remove the battery and boot, device can enter OS and perform
         daily work, and EC does not query the battery info.

Change-Id: Iff8dfc3c64831c0055153f0448de8c45c41896a0