rex: Update battery_is_present logic
Update the battery_is_present() logic to include the battery cut-off
state in determining if the battery is "present".

Modern usage battery_is_present() is more of a logical "battery usable"
check rather than a physical battery presence check. Scenarios like the
ShutdownWithCommandBatteryCutoff test were failing because the caller
assumed it's safe to cut-off the battery when A/C is present.

BUG=b:289566759
TEST=tast power.ShutdownWithCommandBatteryCutoff now passes reliably
	only 1 failure in 490 runs due to "Failed to start Chrome",
	which is not related to the battery.
TEST='./twister -T zephyr/test/rex' passes

Change-Id: Id1f9c98b8652d19d38c0493e5367ccd2e61803e4