common: motion_sense: Fix null pointer dereference
This fixes a very similar problem to the one addressed in:
e8b5147f37bd ("common: motion_sense_fifo: fix iteration bug")
The motionsense logic expects that every data entry if proceeded
by a timestamp entry on the event queue.
The patch mentioned above added some check to ignore entries of wrong
type, or inserted in the wrong order.
The problem is that one edge-case remained unaddressed.
If the very first element of the fifo is of data type we'll try to
fetch the previous, "-1" element. The fifo/queue API will return a NULL,
causing a crash once that pointer is dereferenced.

Fix that by checking if the pointer returned from the fifo is NULL,
and if that's the case simply ignore the current element.
Add a simple test to cover this.

BUG=b:290725559
TEST="suspend_stress_test -c50" doesn't result in EC crash on whiterun

Change-Id: I91b7d1d1fc78d6d8558156a5d655bdfc73fba2be