dts: Add table for zephyr ntc-thermistor driver
The EC and zephyr thermistor drivers read precomputed tables from the
DT. However, the formats of these tables are different. Add a table
compatible with the zephyr driver for one of the thermistors already
defined in thermistor.dtsi.

Note that this table is added as a #define, because all the relevant
properties have to be put directly into the thermistor node, i.e. node
references cannot be used.

BUG=b:291721513
TEST=Example of adding a thermistor device node:
memory-thermistor {
	compatible = "ntc-thermistor-generic";
	io-channels = <&adc0 2>;
	THERMISTOR_3V3_30K9_47K_4050B
};

Change-Id: Ia1c3340e4822b5dd4b74217df17b7d95c1c1a4ec