npcx: Fix GPIO enumeration for CONSOLE_CMD_GPIODBG
This fixes a compilation issue when PLATFORM_EC_CONSOLE_CMD_GPIODBG is
enabled. When IO expanders are defined, the npcx support code fails to
compile because it tries to include definitions for GPIO pins on IO
expanders. These should be filtered out since this command is specific
to npcx on-chip GPIO pins.

TODO: IO expander GPIOs are eroneously included

BUG=b:292291807
TEST='gpiodbg list' works on rex

Change-Id: I89ffd61a213969bf791b6019a97e18d9bdc721fb