ec3po: Port select.select to epoll.
epoll is not as portable as select.select but it is more scalable.

When running a lot of tests - select.select was failing unless the
tests were also threaded.

This CL solves the ec3po - it does not solve all of the problem
with tests as we hit another system limit for pexpect - but that is
not an ec3po issue and will be handled separately.

BUG=b:244182764
TEST=servod starts locally, all tests pass when ran threaded.
TEST=run_tests.sh in ec/util/ec3po runs and passes.

Change-Id: I3ba8c47b8ee473a636c21b5e08d90c7c89250d75