Trogdor: Correct the pin assignment
It uses pin assignment D as 2 lanes are reserved to the USB SS.
Override the get_dp_pin_mode() function to return the value.

The get_dp_pin_mode() is not used to set the TCPC mux and it is
used only by the host command, like the feedback report. This
change fixes the feedback report inform.

BUG=b:292000552
TEST=ectool typecstatus

Change-Id: Ic86e612253e0b151f7a69bc326d0fe7e6630b3b7