npcx/trng: Move trng init logic to startup
Initializing the trng module and generating an initial set of random
numbers can take upwards of 500 ms. Move this logic outside of the
more time critical operations, such as responding to host commands, to
prevent timeouts and failures.

For this to work properly the init logic needs to be called once at
startup and the module must stay powered up. To maintain compatibility
with existing TRNG designs, the trng_init and trng_exit will be
maintained but left empty. We can revisit powering down the TRNG later
if needed to save power.

BUG=b:293651381
TEST=manual testing

Change-Id: I0dbe758e500d57ac6bd1ff8eee6626ef6f484b87