ovis: Change PPC i2c address from 0x40 to 0x42
The PPC address strapping is controlled via control lines. The control
lines aren't drivenn when Ovis is not powered. Hhowever the PPC can
still be powered when attached to a power role source capable port
partner via the PPC's dead battery mode. In this case, the PPC will
latch the i2c device address of 0x42.

This CL adjusts the expected i2c address for the PPCs to be 0x42.

BUG=b:293979153
TEST=Verified on Ovis that the PPCs are initialized successfully using
0x42 device address.

Change-Id: I850c0ba974de5a411d1f320392950e1d0d24882f