watchdog: Add watchdog warning panic reason
Add a new panic reason (PANIC_SW_WATCHDOG_WARN) that is set when a
watchdog warning occurs. This is useful in two ways. First, its a
straight forward way to preserve panic info (e.g. task id) on chipsets
that have a watchdog warning handler but not a watchdog reset handler.
Second, it can be used to collect metrics on watchdog warnings in the
field before they become watchdog resets.

BUG=b:293942684
TEST=Pass new unit tests
TEST=Task id is set after watchdog on skyrim

Change-Id: Ib46ad3a35534af2f8be97f88cee8b7f9c78c1b91