ec/util/comm-dev.cc: Pass sizeof structure during ioctl
The kernel really should be told the sizeof the memory
being passed to it. This probably works out most of the
time because the kernel happens to have a structure of
similar of same size that aligns up however this isn't
guaranteed as it could be larger (not smaller due to
kernel ABI stability rules).

BUG=b:262782150,b:282855423
TEST=none

Change-Id: I66383cde9fed1be5cdac3ee10298593dc98dca6f