nissa: restrict I2C passthrough
Default I2C passthrough policy in the EC is to allow everything, which
can be dangerous because it allows unrestricted access to devices that
may allow code running on the AP to damage the system or cause it to
behave in unexpected ways. Enable the config option to default-deny
passthrough commands from the AP.

Because Nissa devices may use I2C passthrough to update TCPC or retimer
firmware, the policy is customized to allow access to type-C ports
(which normally becomes forbidden dynamically when depthcharge sends a
command to restrict access after updating chip firmwares during normal
boot).

BUG=b:294550904
TEST=twister; zmake build -a

Change-Id: Ie3a9bcd5cca07fbdb4a256ac36f28a5200d6cfed