Karis: Modify GPIOs definition and USBC retimer setting
Move usb_c1_rst_odl from GPIO37 to GPIOC7 and rename to
usb_c1_rt_3p3_sx_en which is the ls-en-pin in C1 HB retimer.
For there is no NCT3807 on Karis, usb_c0_frs_en is moved to
GPIO42 and usb_c0_rt_rst_odl is moved to GPIO37.

The FW_CONFIG bit0-1 is for USB DB field which can switch
the C1 retimer for future used. Karis only use USB4 HB for
C1 now so the other retimers are removed.

BUG=b:296145621
TEST=zmake build karis

Change-Id: I8c4a24a8b201fc1acb7de94dab0de627a6b1eeab