charge_state: Add EC_CMD_CHARGE_CONTROL V3
Currently, only if sustain_soc.lower == sustain_soc.upper, the battery
sustainer uses IDLE (as well as DISCHARGE) for discharging. This forces
a host to set lower < upper to avoid IDLE. This is suboptimal not only
because the behavior is obscure but also because a host can't specify
lower < upper for boards which support IDLE. (That means there is no
hysteresis.)

This CL adds flags field in EC_CMD_CHARGE_CONTROL so that a host can
explicitly disable IDLE.

This CL also adds dedicated tests for EC_CMD_CHARGE_CONTROL to increase
the coverage.

There is no functionality change for the hosts speaking V2.

BUG=b:222619859,b:188457962
TEST=make run-sbs_charging
TEST=ectool chargecontrol normal 79 80 0x1

Change-Id: I9ca2d5f68f9c1d84baff9feb5eefd7b60e98e474