Battery: Redefine BATTERY_CUTOFF_STATE_IN_PROGRESS
Currently BATTERY_CUTOFF_STATE_IN_PROGRESS is set right before a cutoff
request is sent to a battery and immediately changed to CUT_OFF if a
request is accepted or to NORMAL if a request is rejected.
board_cut_off_battery just calls a few I2C commands. Thus this
transitional period represented by IN_PROGRESS isn't very useful.

This CL makes the EC keep NORMAL until a request is accepted and switch
to IN_PROGRESS (if a request is accepted) or NORMAL (if rejected). This
allows us to use IN_PROGRESS to describe the state where a battery is
working on cutoff, which usually takes several seconds.

This CL also allows the EC to transition to CUT_OFF state only after it
verifies cutoff is complete while previously it was unclear.

There is no functionality change from the external perspective.

BUG=b:278898714
TEST=Execute cutoff console command on Eldrid.

Change-Id: I1a482002f97d38a3f02bfd53ee59dfb3f3fe479b