run_device_tests: Explicitly set fp_sensor_type in fpsensor_hw
The fpsensor_hw test can fail due to an incorrect setting of the
'fp_sensor_sel' servo control. Explciitly setting it should not be
necessary but, as discovered in b/295540228, there are cases where it
could be inadvertantly changed. This pre-test hook ensures that it is
set to the correct value.

BUG=b:295540228
TEST=run_device_tests.py --board bloonchipper -t fpsensor_hw
TEST=run_device_tests.py --board dartmonkey -t fpsensor_hw

Change-Id: I591c24676e3651f19adbf8cad5976242e15ed6f3