ec_commands.h: Enable pretty enum names for EC status enum
These will be used in libec - https://crrev.com/c/4779594.

Note that the checkpatch wants no space between
"#define TEXT_MAP_ITEM(k, v)" and "[k] = #v" (BRACKET_SPACE rule), but
clang-format wants the space. This is a bit of a false positive.
We are going to go with clang-format's preference to allow
other to easily invoke clang-format in the future.

BUG=b:294592862
TEST=make buildall

Change-Id: I1cdec6cd48441523c2629c21879744f07bb73f91