zephyr: Initial version of CME (Component Manifest Engine)
Implement the initial version of CME (Component Manifest Engine), which
is a Python program, which extracts the minimal component information
from the Zephyr device tree file and converts them into a component
manifest JSON file.

The CME is executed per each Zephyr project and the component manifest
file is placed to the output directory.

BUG=b:296285094
TEST=zmake build --all

Change-Id: I1071b02818dc8845b91055011dd1992926c2c3c2