usbc: Allow preserving ap-off and stay-in-ro flags across PD reset
power_state:rec with cros_ec_softrec_power works by running `reboot
wait-ext ap-off-in-ro`. On a chromebox which is powered by Type-C only,
the EC reset will result in a PD hard reset and the device will brown
out. When it boots again the ap-off and stay-in-ro flags are lost so
recovery fails.

Add a config to preserve the ap-off and stay-in-ro flags across a PD
hard reset. This should only be enabled on chromeboxes which don't have
servo micro and therefore can't use cros_ec_hardrec_power.

This doesn't affect manual recovery on user devices, since it uses the
recovery signal from the GSC, not the reset flags.

BUG=b:293545949
TEST=With other workarounds, FAFT passes on boxy powered by Type-C
(results in b:271570963).

Change-Id: Id9674a004d332d4d3a439bafd9304b945682172d