opentitantool: add opentitantool wrapper script
It is quite difficult to have portage build opentitantool using ebuilds,
but opentitantool can build easily inside and outside of the chroot
using bazel.

Add a wrapper script that ensures that opentitantool is built and
forward all command line arguments to the built binary.

We will install a syslink to this script in chroot /usr/sbin/ to make it
even easier for people to execute opentitantool commands

BUG=b:294848143
TEST=run ./utils/opentitantool.sh from a few locations and see that is
forward  command correctly

Change-Id: Ib1cfd8524c54dc14cf0c313a93874ce7a4588c96