board/hyperdebug: Allow runtime reconfiguration of chip select
HyperDebug is a versatile device, operating in a similar way to uServo.
It makes several SPI ports abaiable for the host computer to control.

This CL adds a new console command for changing which pin is to be used
as the chip select signal, allowing HyperDebug to control multiple SPI
devices sitting on the same bus (that is, sharing data and clock
signals).

BUG=none
TEST=logic analyzer of manual runs with modified opentitantool

Change-Id: I38b56c5bf57c8858b67c867e0a4625d5134ac16b