zephyr: config: drop few unnecessary features at application level
Drop a couple of shell commands and a flash API. The watchdog one is new
upstream and enabled by default there, the flash ones are commonly
disabled in few boards to save some flash. Just switched them off by
default and simplify the configs a bit, they can be enabled for
debugging.

BUG=none
TEST=cq dry run

Change-Id: If0561ae2d4f0efb6afd339b708d9d75f7a619d66