emul_power_signals: disable unneeded signal types
When using power signals emulation without eSPI enabled (such as if all
required power signals are GPIO-based), the code to handle emulated eSPI
signals would fail to build because no eSPI support exists. Add
preprocessor guards to remove this support to fix the build in these
configurations.

This also drops inclusion of test_state.h, which is unused in this file.

BUG=b:271118604
TEST=twister

Change-Id: I29dd57a25413feba447e63c435699b6c065d6d79