fpsensor: Add states for the fingerprint templates
In go/eliminate-fpmcu-preload, we proposed three new states to the
template data.
1. Encrypted template
2. Unlocked decrypted template
3. Locked decrypted template

The state will be used to track the status of template, and the template
can only be used for matching when the state is unlocked.
For the detail state transitions, please reference the design doc.

And add EC_MKBP_FP_ERR_MATCH_NO_AUTH_FAIL to distinguish the
authentication error in the matching operation.

BUG=b:312339856
TEST=make runhosttests -j200 >/dev/null
LOW_COVERAGE_REASON=b:303755024 no test coverage for fpsensor.cc. And
                    there is now way to cover fp_process_match now.

Change-Id: I4648c7cc6110629b32458ad7a6c26467bbcf9e2f