zephyr: test: make emul based tests work with init priority checking
Build time init priority checking is now default upstream so it runs by
default in the tests as well.

Make this work with the tests by lowering the priority of the busses
with emulators on the them to the priority of the emulator itself.

For few emulator drivers this meant that they started initializing
before the driver itself, which was causing issues when resetting some
interrupt lines that have not been configured as input yet, fix that by
not resetting the line on init (it's already at the initial state
anyway) but keep resetting it in between tests.

For one of the driver tests, we were hitting a hardcoded build check
against the default kernel init priority, let's leave the option off
there.

Long term it may be worth adding an option to tweak the emulator init
priority, but this should close the gap for now.

BUG=b:275577512
TEST=cq dry run

Cq-Depend: chromium:4809907
Change-Id: Ia70811a22265ac1c8ffeaa74b82730c65abc42de