util/flash_ec: Handle missing CBI region
It's better to scan the original flash image to check if any existing
regions need to be preserved rather than scanning the new flash
image. This way, we won't fail trying to preserve a non-existent region
when we introduce new preserved regions.

If an existing preserved region is no longer present in the new image,
flash_ec needs to be invoked with --no_preserve to effectively remove
the preserved region.

BUG=b:286288547
TEST=flashed rex with new EC image that transitions from no CBI to CBI

Change-Id: I92bced665bed2c0ac5fe975dd53a695262e9d41b