zephyr: guard "only one" assertions against zero
Some drivers support only one instance of a compatible in the device tree
and statically assert that no more than one exists, but fail to check
whether there are actually zero instances, which is allowed. Add a check
for zero instances to avoid saying "only one" item is allowed if there
are instead none.

BUG=none
TEST=twister

Change-Id: Ic96c5ba446736f87124a50cc58fa53f9a52bef2f