zephyr: bloonchipper: add support for host commands
Bloonchipper uses can use two interfaces to communicate with host - UART
and SPI. It is detected in runtime which interface should be used, based
on the GPIO state.

Add initialization of a proper host command backend in runtime. Also add
handler for the EC_CMD_GET_PROTOCOL_INFO, which also returns values
appropriate for a present interface.

BUG=b:259667397
TEST=zmake build bloonchipper && enable commands and run 'ectool
--name=cros_fp stress'
LOW_COVERAGE_REASON=limitation of the GPIO emulator - no possibility of
returning an error

Change-Id: I93713e893b60228e73d2cd35ba7d894412faf13f