driver: bmi: Avoid infinite loops in the irq_handler
The BMI accelgryo drivers all have loops in their bottom-half
interrupt handler. These loops run as long as an interrupt continues
to be reported by the hardware.

This is, apparently, needed for BMI hardware because we're using them
in "edge triggered" mode. It is believed that the edge won't
necessarily be asserted again unless we fully clear all interrupt
sources so that the interrupt line can go high again and then see
another edge.

Unfortunately, it's suspected that in the BMI accelgryo drivers (or
hardware) there is some bug (unknown what) that causes us to end up in
an infinite loop state. Specifically, we've seen field reports on
sc7180-trogdor-lazor where the motion sense task simply stops running
even though sensors are enabled. Code inspection makes us believe that
the BMI interrupt handler might be looping.

Let's cap the number of times that we loop in the BMI accelgryo
drivers at 200 times. There's nothing magic about looping 200 times,
but we expect this to be far and above more than enough times in
reality. A few notes:

* On BMI160 the hardware FIFO is 1024 bytes. On BMI260 it's 2048 bytes
  (though, oddly the FIFO length register allows for expressing 16K
  since it's 14 bits). On BMI3xx it's 2048 bytes.

* Each time through the loop we read up to 64 bytes of data
  (BMI160/260 as per BMI_FIFO_BUFFER) or 32 bytes of data (BMI3xx as
  per BMI3_FIFO_BUFFER). We might read less than this if there was
  more than one frame and later frames are cut off. As a simplifying
  assumption let's say we read 32 bytes of data per loop on BMI160 and
  16 on BMI3xx. That would mean in 200 loops we'd read at least 3200
  bytes of data.

* Any time there's more data in the HW FIFO than our buffer size, the
  driver currently prints out "unexpected large FIFO". Given that EC
  logs are not filled with this message, it appears uncommon to
  actually hit the case where we need to loop in order to read all the
  data out of the FIFO.

In the case that we really do hit the 200 loops, we'll print a message
and attempt to flush the FIFO. Hopefully this will bring the device
back to some semblance of normalcy.

As part of this, we also fix it so that if we process at least one
interrupt we'll still call motion_sense_fifo_commit_data() and we'll
also still return success from the interrupt handler saying that we
got some data. It's unlikely this case is being hit, but it seems
(from code inspection) like a slightly better way to deal with that
error case.

In case it's not obvious from the above, we don't know if this will
fix any problems in the field. However, it seems highly unlikely to
_cause_ any problems.

BUG=b:267680317
BRANCH=trogdor
TEST=Use motion sense on lazor with BMI160
Unit tests.

Conflicts: No unit tests for sensors in branch.
	zephyr/test/drivers/bmi3xx/src/bmi3xx.c
	zephyr/test/drivers/default/src/bmi160.c
	zephyr/test/drivers/default/src/bmi260.c

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4818859
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ia35fd123313ac6681aee50661f88394dc770de41
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>