driver: bmi: Avoid infinite loops in the irq_handler
The BMI accelgryo drivers all have loops in their bottom-half
interrupt handler. These loops rrun as long as an interrupt continues
to be reported by the hardware.

Loops in interrupt handlers like this aren't unheard of. If there's
already a new interrupt pending by the time you finished processing
the previous one it can make sense to process it right away to avoid
the overhead of processing another interrupt. On the flip side,
however, if we _always_ loop when an interrupt is present then it has
the possibility that a bug in the driver could lead to an infinite
loop.

It's suspected that in the BMI accelgryo drivers there is some bug
(unknown what) that causes us to end up in the infinite loop
state. Specifically, we've seen field reports on sc7180-trogdor-lazor
where the motion sense task simply stops running even though sensors
are enabled. Code inspection makes us believe that the BMI interrupt
handler is likely looping.

Let's cap the number of times that we loop in the BMI accelgryo
drivers at 3. In theory, looping twice would probably have been
enough. The mostly likely time for us to get two interrupts in a row
would be if we took a long time getting to the interrupt handler to
process the first one. However, we'll do 3 just to be safe.

In practice, normal usage (including running an Android app that
reports accel/gyro state) never saw the loop trigger on my lazor. We
always exited with i == 1 && rv == 0.

As part of this, we also fix it so that if we process at least one
interrupt we'll still call motion_sense_fifo_commit_data() and we'll
also still return success from the interrupt handler saying that we
got some data.

BRANCH=None
BUG=b:267680317
TEST=Use motion sense on lazor with BMI160

Change-Id: Ia35fd123313ac6681aee50661f88394dc770de41
Signed-off-by: Douglas Anderson <dianders@chromium.org>