watchdog: Always set exception number to current task
By convention, the task id is set in the panic exception number
field when a watchdog occurs. When the watchdog occurs while in
an ISR/handler context the task id is set to 0xff. This makes
determining which task is misbehaving difficult. Instead always
set the current task id, which will match the most recent task
if inside an ISR.

BUG=None
TEST=Task id matches expected task after watchdog

Change-Id: Id76366e956c9774bf96cd86d2b0df06bdef1824d