command: Add preserve_old_hostcmd_flag param to EC_CMD_GET_PANIC_INFO
EC_CMD_GET_PANIC_INFO sets the PANIC_DATA_FLAG_OLD_HOSTCMD flag.
The crash collector checks this flag to determine if the panic data
is stale or not. This means reading the panic info before the crash
collector will result in no crash report being generated.

This change adds an option to preserve the PANIC_DATA_FLAG_OLD_HOSTCMD
flag when reading the panic info.

This option is only available in >= version 1 of the command.

ectool is updated to preserve this flag when available.

BUG=b:273967724
TEST=Waitms 1400 to set watchdog warning panic info, repeated calls
to ectool panicinfo report (NEW)

Change-Id: Ib254a051abc2eafe64f8f10b6c42529fd75cbd1a