keyboard: Make refresh key work as boot key
This CL makes a refresh key work as a boot key. It adds a hook for power
button release (or press). When the PB is released, the EC reads the
refresh key state. This is done only for the first PB release.

BUG=b:285187778, b:299186434
TEST=Press Refresh+Power to reset Eldrid and release the power button at
different timings. Verify the refresh key is consistently detected as a
boot key.
TEST=Eldrid can enter recovery mode by Esc+Power+Refresh.
TEST=./twister -s drivers/drivers.keyboard_scan

Change-Id: I9062b025735cabbbef226c4f222e381592366a1c