ServoV4p1: Disable virtual EMCA emulation by default
ServoV4p1 has added command to support eMarked Cable Emulation (EMCA).
This is present on all ServoV4px, but was made default when adopted.

This CL changes default back to not emulating an eMarked cable (in SNK
and SRC, non-DTS states). The reasons are twofold:

(1) This change was not backported to v4p0 due to flash limits.
    Leading to inadvertent fork in functional behavior between the two.

(2) There is an issue with VCONN straining the PP3000 rail that may be
    alleviated by disabling VCONN delivery as much as possible*.

    (*Note: Per spec hosts *may* drive CC2/Vconn blindly without, vRa.)

This is a minor change that should return v4p1 to v4p0 state. Full
eMarker SOP' TCPM emulation is left as TODO (b:171291442).

BUG=b:149563547
BRANCH=main
TEST=1. Build firmware
     2. Flash ServoV4p1 with DRP toggling DUT
     3. Run `cc snk` after reboot
     4. Monitor cc line for Servo emarker emulation

Change-Id: I1b7c64899b3981b7729975a895d1c2e70dfc614e