chip/stm32: Allow per-SPI-port advertising of full duplex
The USB_SPI_CUSTOM_SPI_DEVICE flag in spi_device_t.usb_flags allows
board to declare that a board-specific driver should be used instead of
the common STM32 spi host driver, for a particular SPI port.  This CL
further allows board to declare whether the custom driver supports full
duplex or not (independently of whether the common STM32 spi driver is
compiled with full duplex support or not.)

This is of particular use on the HyperDebug board, which has a Quad-SPI
port that is programmed quite differently from the other SPI ports, and
among other differences, does not support full duplex.

BUG=b:299351632
TEST=make BOARD=hyperdebug

Change-Id: Ide3b2ceb92da4ae549fe9f511d0b523d4ab7787d