gstool: add --wp [follow|disable] options for Ti50
The default wp protect state for ti50 will be force enable in the
future. We also want a hook to inform users that disabling or making
write protect follow the chassis should require AllowUnverifiedRo to be
set to always otherwise modifications to AP RO section could soft brick
a device and require a SuzyQ to fix it.

See corresponding ti50 change: CL:*6443638

BUG=b:258288714,b:257255419
TEST=ensure that disable and follow require AllowUnverifiedRO to be
always
TEST=ensure that using new commands with cr50 error saying that this
only work for Ti50

Change-Id: If9bebe5084a2a2a05b2ac98a731f49eef4f3d7b3