zephyr: Fix snprintf format string
system_get_chip_version will return a value from 0 to 15.
cros_system_it8xxx2_get_chip_revision adds 0xa to that resulting in a
value from 0xa to 0x19, so the printf %1x will print a value that is 1
or 2 chars. It appears that what we really want to do is convert the
revision from a number to a letter, i,e 0=A, 1=B, etc.

Switch to printing an ascii char from a to p instead of hex a - 19.
Sadly there are no unit tests for this code.

BUG=None
TEST=zmake build geralt
LOW_COVERAGE_REASON=This will unblock gitlab failures, b/299640211

Change-Id: Ibc4654ede560bc1142f1beb9433f36f2080d9c22