chip/stm32: Avoid gpio glitch when leaving open drain mode
The code already considers, that when applying GPIO_ODL_HIGH, it is
important to set the open drain flag BEFORE enabling output drive, or
else, the pin will be strongly driven high for a short time.

However, the existing code did not adequately handle the case where such
a pin is afterwards converted to e.g. GPIO_INPUT.  And prior to this CL,
a switch from open drain to input would lead to a brief period of
actively driving high.

BUG=b:299669075
TEST=observe glitch-free switch with logic analyzer

Change-Id: I9c7f0d4709c8d14b29074d0c79c72481f73774e9