chip/stm32: Add flexibility to USB stream interface descriptor
USB stream interfaces have two endpoints, an IN and an OUT endpoint.
Usually, the order in which they are listed in the USB descriptor is not
important, as only the endpoint numbers, such as 0x01 and 0x81 are used.

However, the CMSIS-DAP protocol specifies that the OUT endpoint must
come first, and the IN endpoint as the second, in the descriptor.  This
could be achieved by simply reversing the order in the
USB_STREAM_CONFIG_FULL macro.  However, it is impossible to know if some
utility program (an alternative to flashrom, maybe), would choke on the
descriptors being reversed.  Because of that, I do not want to flip them
all.  Instead, this CL enhances the USB_STREAM_CONFIG_FULL macro with
additional parameters, such that it can generate the descriptors in any
order. All existing call sites are updated in such a way that there will
be no change in actual descriptors.

BUG=b:299669075
TEST=make BOARD=hyperdebug

Change-Id: Ia6801008476a0928129921816895de6e16386987