nx20p3483: fix ppc overvoltage protection on brya board
The overvoltage protection threshold was not set properly.
After enabling the sink path, the value of threshold register
was set to the highest possible value. Since the hardcoded value
for the driver was also the highest possible, it's not a major
danger, but in case the value should be lower, it wouldn't be set
properly making the overvoltage protection do not work as desired.

BUG=b:299913813
TEST=compile the CL: for brya board
     connect the charger through voltage meter and see that despite
     the threshold being set to 10V, the current is flowing
     while having the 20V from charger. After applying this CL,
     the current is properly stopped from flowing. Changing the
     requested voltage to values below threshold makes the current
     flow correctly.
TEST=Execute the "ppc 1" shell command and verify the value of
     threshold register at offset 8:
     [0x08]: 0x02

Change-Id: I7d9744739163d5602a55539c68dcf60f8c0d37d2