panic_output: On assert, set top two info bytes to file name
On assert, the panic info register is set to the line number. Inferring
the exact assert from just the line number is cumbersome. Improve this
process by setting the top two bytes of the info register to the first
two characters of the file name. This will significantly speed up
debugging assert panics.

BUG=b:297382155,b:295773669
TEST=Pass drivers.default twister tests

Change-Id: I5f2ce652a986a0e0b4d8705a6e2299e721ab879a