charge_state: Remove special system-not-locked case
The original code computed a different input charging current when the
system is locked vs. not locked. This resulted in the potential for an
unreasonably high current limit in unlocked systems that could over-draw
chargers with lower ratings than the system's maximum current draw
rating. For example, if the system is rated at 95 W, 5 A but a 65 W,
3.25 A charger is connected, we would significantly over-draw the
charger.

I can't think of a case where we would want to allow this violation in
unlocked (pre-MP) systems, while properly handling MP systems. so,
remove this special case.

before this change:

ec:~> chgsup

Port --Supplier-- Prio -Available Power-
*P0  PD             1   3250mA 20000mV
     USBC_U1_5A     4    500mA  5000mV
     VBUS           4    500mA  5000mV

  Left safe mode
  Override port = P-1

ec:~> charger
  Name:   ISL9241
  Option: (0x1030004)
  Man id: 0x0049
  Dev id: 0x000e
  I_dptf: disabled
Limits			 ( min     max step)
  chg_voltage:  12984 mV (  64 - 18304,   8)
  chg_current:      0 mA (   8 -  6140,   8)
  input_current: 3080 mA (   8 -  6140,   8)
ec:~>
ec:~> cutoff
[64.172900 Battery cut off]
ec:~>
ec:~> charger
  Name:   ISL9241
  Option: (0x1030004)
  Man id: 0x0049
  Dev id: 0x000e
  I_dptf: disabled
Limits			 ( min     max step)
  chg_voltage:  12984 mV (  64 - 18304,   8)
  chg_current:      0 mA (   8 -  6140,   8)
  input_current: 4744 mA (   8 -  6140,   8)
ec:~>

note how we're setting the input_current to 4.75 A for a 3.25 A charger
(derated by 5% to 3.08 A) in response to a battery cut-off.

ec:~> charger
  Name:   ISL9241
  Option: (0x1030004)
  Man id: 0x0049
  Dev id: 0x000e
  I_dptf: disabled
Limits                     ( min     max step)
  chg_voltage:  13048 mV (  64 - 18304,   8)
  chg_current:   3920 mA (   8 -  6140,   8)
  input_current: 3080 mA (   8 -  6140,   8)
ec:~>
ec:~> cutoff
[36.813300 Battery cut off]
ec:~>
ec:~> charger
  Name:   ISL9241
  Option: (0x1030004)
  Man id: 0x0049
  Dev id: 0x000e
  I_dptf: disabled
Limits                     ( min     max step)
  chg_voltage:  13032 mV (  64 - 18304,   8)
  chg_current:      0 mA (   8 -  6140,   8)
  input_current: 3248 mA (   8 -  6140,   8)
ec:~>

after this change, we set the limit to 3.25 A. The 5% derating is not
applied in this code path - that's a bug for another day.

BUG=b:296952944
TEST=boots on rex

Change-Id: I0a10a0439d8384ac7fab11044778bafbd6385c9a