extra: i2c_pseudo: reimplement with ioctl()-based UAPI
The i2c-pseudo module provides /dev/i2c-pseudo interface for
implementing I2C adapters in userspace, renamed from
/dev/i2c-pseudo-controller in the original implementation.

This new implementation has an ioctl() interface. The original
implementation, which is replaced in this change, had an
interface of ASCII commands used with read() and write().

The new i2c-pseudo is intended to be more acceptable for upstream
consideration. This change however is not an upstream patch, this
just replaces the existing out-of-tree version.

BUG=b:275615681
TEST=Tested using apel DUT + servo_v4 + ccd_cr50

Servod was built with CL:4541281.

0) DUT has EC SW sync disabled
1) started servod
2) verified expected EC FW version on EC console
3) read EC FW, resulting FW file matched FW file flashed previously via Servo
4) flashed EC FW to different version
5) verified expected EC FW version on EC console (different from step #2 version)
6) read EC FW, resulting FW file matched file flashed in step #4
7) flashed EC FW back to starting version (using step #3 file)
8) verified expected EC FW version on EC console (same as step #2 version)
8) read ITE EC FW, resulting FW file was correct (identical to step #3 file)

There were no errors from i2c-pseudo or servod during this testing.

Also tested manual I2C transactions using i2c-tools, such as:
$ cros_sdk dut-control -o servo_v4_i2c_pseudo_adapter_num
12
$ i2ctransfer -y 12 w1@0x20 0x00 r1
0xa8

Change-Id: I625b9dbf2c117fc1692555a0109308e0a5f1e211
Signed-off-by: Matthew Blecker <matthewb@chromium.org>