ecusb: Enable firmware updates with oooold firmware
Very old servo v4 firmware fails to work with servo_updater, although it
updates when running servo_updater through strace. One suspicion was
that the serial console is too slow, but it wasn't the transfers
themselves, just the initialization of the connection.

BUG=b:227509479
TEST=Updated from servo_v4_v1.1.5734-4a47178 to current. Failed without
this patch (stuck at "Waiting for USB device" until timeout), works with
this patch. If the sleep is one line earlier (before the open), it also
fails.

Change-Id: Ie2430ac78a99ff3dce3af1522ae14fc44c8f0c39