servo_updater: Provide a way to prevent rollback
So far, the updater only checked for equality of version strings and
overwrote the device's firmware with the candidate as soon as there
was any difference.

With this change, it's possible to tell servo_updater using
--no-allow-rollback to skip writing the firmware if the candidate's
version is older than the installed one.  Once fleet (as major user
of servo_updater) has verified that its processes --force writes as
appropriate (if there are any situations where they want to write the
candidate firmware, no questions asked), this option's default can
be toggled to only decrease the installed version number when desired.

BUG=b:182015929
TEST=modified the version strings in the code, checked that the behavior
is as expected with --allow-rollback, --no-allow-rollback, no such
option, for version differentials smaller, equal or larger.

Change-Id: I375f5c785ff274380e5cf78ec87ea915986089ae